[CWS_P2P] prevent spurious WIFI_P2P_STATE_CHANGED_ACTION intent to be broadcasted
Spurious WIFI_P2P_STATE_CHANGED_ACTION intent with DISCONNECT or CONNECT state
are sent when transitioning from UserAuthorizingJoinState to GroupCreatedState
(so when a new device try to join an already created group).
Depending on how application handle this intent, some unexpected behavior
could occurs.
For example with screen cast app, when a device is connected to an adapter and an
other device try to join the group. If the User decline the invitation, the
screen cast session will be lost because a DISCONNECT state is sent with
WIFI_P2P_STATE_CHANGED_ACTION intent.

This is due to current Wifi P2p state machine tree hierarchy:
     \
  EnabledState
       \
    GroupCreatedState
        \
   UserAuthorizingJoinState

As explained in StateMachine.java : "... We then exit from the current state and its
parent's up to but not including the common parent state and then enter all
of the new states below the common parent down to the destination state"

So when transitioning from UserAuthorizingJoinState to GroupCreatedState, the
common ancestor is EnabledState, so the exit/enter calls are:
UserAuthorizingJoinState / exit
GroupCreatedState / exit  => cause intent broadcating with DISCONNECTED
GroupCreatedState / enter => cause intent broadcating with CONNECTED

Modifying the tree state like below with a passthrough state and replacing all
transitions to GroupCreatedState by a transition to this new state permits to
keep the same behavior regarding p2p, but avoid unexpected calls to
GroupCreatedState enter/exit.

                    \
                 EnabledState
                        \
                   GroupCreatedState
                    /             \
GroupCreatedFinalState             UserAuthorizingJoinState

And the WIFI_P2P_STATE_CHANGED intent with DISCONNECT or CONNECT state
is only broadcasted when necessary.

Change-Id: I7615e3b3b19305cdb163ba3b2c7cc406fc3ed29f
Category: aosp improvement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-2342
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: https://android.intel.com:443/265387