[CWS_WIDI_UX] Delete p2p group when forgetting display
When a wifidisplay is forgotten in WifiDisplay Settings activity,
the coresponding p2p group is not removed.

In order to delete the p2p group associated to the wifidisplay, the
function that updates the list of group in WifiP2pService has been
modified so that it also updates the client list for each group.
Thus when a display is forgotten, WifiDisplayController retrieves
the group list from  WifiP2pService, finds the group in which the
display was associated and deletes it.

Change-Id: I0290742a85f4f07306ab337bbd8e6ea0ae695cdf
Category: aosp improvement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-11392
Signed-off-by: Cedric IZOARD <cedricx.izoard@intel.com>
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Reviewed-on: https://android.intel.com:443/307454