WLAN: Handle card removal/insertion for ICC secured networks
For security reasons it's better to disconnect and disable the WiFi networks
that use security related to ICC (EAP-SIM or EAP-AKA) when the user
removes the SIM card.
This is to prevent abusive use of the same SIM card for authenticating
multiple devices to an access point.
Wen SIM card is inserted back, this networks are enabled back too.

Change-Id: I2d8f5621bd68a623b294601d2eaf9bca41917586
Category: feature differentiation
Domain: CWS.WIFI-EAP_SIM_AKA
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-13528
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Signed-off-by: Abdelmajid MLAYEH <abdelmajidx.mlayeh@intel.com>
Reviewed-on: https://android.intel.com:443/308401
Reviewed-on: https://android.intel.com:443/311898